# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Song {
  id: string;
  name: string;
  artists: string[];
  albumName: string;
  image: string;
  language: string;
  duration: number;
  /** @nullable */
  year?: string | null;
}

export interface SongDetail {
  id: string;
  name: string;
  artists: string[];
  albumName: string;
  image: string;
  language: string;
  duration: number;
  /** @nullable */
  year?: string | null;
  streamUrl: string;
  /** @nullable */
  lyrics?: string | null;
}

export interface SongListResponse {
  songs: Song[];
  total: number;
}

export interface Album {
  id: string;
  name: string;
  artists: string[];
  image: string;
  language: string;
  year: string;
  songCount: number;
}

export interface AlbumDetail {
  id: string;
  name: string;
  artists: string[];
  image: string;
  language: string;
  year: string;
  /** @nullable */
  description?: string | null;
  songs: Song[];
}

export interface AlbumListResponse {
  albums: Album[];
}

export interface Artist {
  id: string;
  name: string;
  image: string;
  /** @nullable */
  role?: string | null;
}

export interface ArtistDetail {
  id: string;
  name: string;
  image: string;
  /** @nullable */
  bio?: string | null;
  songs: Song[];
  albums: Album[];
}

export interface GlobalSearchResult {
  songs: Song[];
  albums: Album[];
  artists: Artist[];
}

export interface Chart {
  id: string;
  name: string;
  image: string;
  language: string;
  songCount: number;
}

export interface ChartsResponse {
  charts: Chart[];
}

export type SearchSongsParams = {
q: string;
language?: string;
page?: number;
limit?: number;
};

export type GetTrendingSongsParams = {
language?: string;
limit?: number;
};

export type GetTrendingAlbumsParams = {
language?: string;
limit?: number;
};

export type GlobalSearchParams = {
q: string;
};

export type GetChartsParams = {
language?: string;
};