# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: string;
  name: string;
  artists: string[];
  image: string;
  language: string;
  year: string;
  songCount: number;
}