# lib/api-zod/src/generated/types/albumDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface AlbumDetail {
  id: string;
  name: string;
  artists: string[];
  image: string;
  language: string;
  year: string;
  /** @nullable */
  description?: string | null;
  songs: Song[];
}