# lib/api-zod/src/generated/types/albumListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */
import type { Album } from './album';

export interface AlbumListResponse {
  albums: Album[];
}