# lib/api-zod/src/generated/types/artist.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */

export interface Artist {
  id: string;
  name: string;
  image: string;
  /** @nullable */
  role?: string | null;
}