# lib/api-zod/src/generated/types/artistDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */
import type { Album } from './album';
import type { Song } from './song';

export interface ArtistDetail {
  id: string;
  name: string;
  image: string;
  /** @nullable */
  bio?: string | null;
  songs: Song[];
  albums: Album[];
}