# lib/api-zod/src/generated/types/chart.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */

export interface Chart {
  id: string;
  name: string;
  image: string;
  language: string;
  songCount: number;
}