# lib/api-zod/src/generated/types/songDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */

export interface SongDetail {
  id: string;
  name: string;
  artists: string[];
  albumName: string;
  image: string;
  language: string;
  duration: number;
  /** @nullable */
  year?: string | null;
  streamUrl: string;
  /** @nullable */
  lyrics?: string | null;
}