# lib/api-zod/src/generated/types/songListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * iMusic Indian Cinema Songs API
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface SongListResponse {
  songs: Song[];
  total: number;
}